requirements.txt
requests
markdownify
beautifulsoup4
pypdf
chromadb
cohere
python-dotenv
langchain_text_splitters
ipywidgets
fastapi
uvicorn